Move QR scanner to new file
